[generator] Replace wiki link with a shorter version
