+++ android/src/com/mapswithme/maps/widget/placepage/PlacePageView.java
@@ -64,7 +64,7 @@ import com.mapswithme.util.statistics.AlohaHelper;
 import com.mapswithme.util.statistics.Statistics;
 
 import java.io.UnsupportedEncodingException;
-import java.net.URLDecoder;
+import java.net.URLEncoder;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -427,13 +427,7 @@ public class PlacePageView extends RelativeLayout implements View.OnClickListene
     refreshMetadataOrHide(mMapObject.getMetadata(Metadata.MetadataType.FMD_EMAIL), mEmail, mTvEmail);
     refreshMetadataOrHide(mMapObject.getMetadata(Metadata.MetadataType.FMD_OPERATOR), mOperator, mTvOperator);
     refreshMetadataOrHide(translateCuisine(mMapObject.getMetadata(Metadata.MetadataType.FMD_CUISINE)), mCuisine, mTvCuisine);
-    try
-    {
-      final String wikipedia = mMapObject.getMetadata(Metadata.MetadataType.FMD_WIKIPEDIA);
-      refreshMetadataOrHide(TextUtils.isEmpty(wikipedia) ? null : URLDecoder.decode(wikipedia, "UTF-8"), mWiki, mTvWiki);
-    } catch (UnsupportedEncodingException e)
-    {
-    }
+    refreshMetadataOrHide(mMapObject.getMetadata(Metadata.MetadataType.FMD_WIKIPEDIA), mWiki, mTvWiki);
     refreshMetadataOrHide(mMapObject.getMetadata(Metadata.MetadataType.FMD_INTERNET), mWifi, null);
     refreshMetadataOrHide(mMapObject.getMetadata(Metadata.MetadataType.FMD_FLATS), mEntrance, mTvEntrance);
     // TODO throw away parsing hack when data will be parsed correctly in core
@@ -699,8 +693,14 @@ public class PlacePageView extends RelativeLayout implements View.OnClickListene
       break;
     case R.id.ll__place_wiki:
       final String[] wikiParts = mTvWiki.getText().toString().split(":");
-      if (wikiParts.length == 2)
-        followUrl("https://" + wikiParts[0] + ".wikipedia.org/wiki/" + wikiParts[1]);
+      try
+      {
+        if (wikiParts.length == 2)
+          followUrl("https://" + wikiParts[0] + ".wikipedia.org/wiki/" + URLEncoder.encode(wikiParts[1].replace(' ', '_'), "UTF-8"));
+      }
+      catch (UnsupportedEncodingException e)
+      {
+      }
       break;
     case R.id.tv__bookmark_group:
       saveBookmarkNameIfUpdated(null);
