+++ generator/osm2meta.hpp
@@ -213,51 +213,6 @@ protected:
   {
     return v;
   }
-
-  // Special URL encoding for wikipedia:
-  // Replaces special characters with %HH codes
-  // And spaces with underscores.
-  string WikiUrlEncode(string const & value) const
-  {
-    ostringstream escaped;
-    escaped.fill('0');
-    escaped << hex;
-
-    for (auto const & c : value)
-    {
-      if (isalnum(c) || c == '-' || c == '_' || c == '.' || c == '~')
-        escaped << c;
-      else if (c == ' ')
-        escaped << '_';
-      else
-        escaped << '%' << std::uppercase << setw(2) << static_cast<int>(static_cast<unsigned char>(c));
-    }
-
-    return escaped.str();
-  }
-
-  string ValidateAndFormat_wikipedia(string const & v) const
-  {
-    // Find prefix before ':', shortest case: "lg:aa".
-    string::size_type i = v.find(':');
-    if (i == string::npos || i < 2 || i + 2 > v.length())
-      return string();
-
-    // URL encode lang:title (lang is at most 3 chars), so URL can be reconstructed faster.
-    if (i <= 3 || v.substr(0, i) == "be-x-old")
-      return v.substr(0, i + 1) + WikiUrlEncode(v.substr(i + 1));
-
-    static string::size_type const minUrlPartLength = string("//be.wikipedia.org/wiki/AB").length();
-    if (v[i+1] == '/' && i + minUrlPartLength < v.length())
-    {
-      // Convert URL to "lang:title".
-      i += 3; // skip "://"
-      string::size_type const j = v.find('.', i + 1);
-      static string const wikiUrlPart = ".wikipedia.org/wiki/";
-      if (j != string::npos && v.substr(j, wikiUrlPart.length()) == wikiUrlPart)
-        return v.substr(i, j - i) + ":" + v.substr(j + wikiUrlPart.length());
-    }
-    return string();
-  }
+  string ValidateAndFormat_wikipedia(string const & v) const;
 
 };
